Turn `quill::plugin` into a proc macro
