+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
     "quill/sys",
     "quill/common",
     "quill/api",
+    "quill/api/plugin-macro",
     "quill/plugin-format",
     "quill/cargo-quill",
