+++ quill/api/Cargo.toml
@@ -5,6 +5,7 @@ authors = ["caelunshun <caelunshun@gmail.com>"]
 edition = "2018"
 
 [dependencies]
+plugin-macro = { path = "./plugin-macro" }
 libcraft-core = { path = "../../libcraft/core" }
 libcraft-particles = { path = "../../libcraft/particles" }
 libcraft-blocks = { path = "../../libcraft/blocks" }
