+++ quill/api/plugin-macro/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "plugin-macro"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "1.0.86", features = ["full"] }
+quote = "1.0.14"
